Arrangements after partial delivery
-Clean Code (Organization)
    -Redux
    -SCSS for each component
-View Event
    -Print Date: Month DD of YYYY
-Create Event
    -Disabled button when created
    -Color Select
-Your Events
    -Fix checkbox
