+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Components/ViewEvent/ViewEvent.scss
@@ -1,0 +1,11 @@
+.App-rowButtons-viewEvent{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+}
+
+.App-button-view{
+    background-color: white;
+    margin-top: 18px;
+    margin-left: 4px;
+}
