+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Const.js
@@ -1,0 +1,4 @@
+//Redux
+const CHANGE_TAB = 'change_tab';
+
+export default CHANGE_TAB;
