+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Google/SearchBar.js
@@ -33,5 +33,5 @@ class SearchBar extends React.Component {
 }
 
 export default GoogleApiWrapper({
-    apiKey: ("AIzaSyBdw9I1llo_EUaw_8Wdmsbo_8NvJ2ch98k")
+    apiKey: ("google_key")
 })(SearchBar)
