+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Loader/Loader.scss
@@ -1,0 +1,15 @@
+.App-loader-container{
+    height: 100px;
+    width: 100px;
+    background-color: white;
+    border-radius: 25px;
+    position: fixed;
+    top: 40%;
+    left: 50%;
+  
+    .App-loader{
+      position: relative;
+      left: 10px;
+      top: 10px;
+    }
+}
