+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Redux/Actions.js
@@ -1,0 +1,11 @@
+import CHANGE_TAB from '../Const';
+
+//Action
+let changeTab = function(actualTab) { 
+    return {
+      type: CHANGE_TAB,
+      payload: actualTab
+    }
+}
+
+export default changeTab;
