+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Redux/ReduxNavigationContainer.scss
@@ -1,0 +1,183 @@
+.App-container {
+    display: flex;
+    justify-content: flex-start;
+    margin-left: 10%;
+    margin-top: 3vh;
+    max-width: 80%;
+    border: 0.1em solid #E6E6E6;
+    background-color: white;
+    flex-direction: column;
+  
+    .App-row-elements{
+      flex-direction: row;
+      margin: 12px 0px 12px 18px;
+  
+      .image-left{
+        @media (max-width: 375px) {
+          width: 280px;
+          height: 200px;
+        }
+        @media (max-width: 320px) {
+          width: 230px;
+          height: 180px;
+        }
+        width: 320px;
+        height: 220px;
+        float: left;
+      }
+  
+      .image-event{
+        border-radius: 25px;
+        width: 580px;
+        height: 420px;
+        margin-right: 8px;
+        margin-bottom: 24px;
+      }
+  
+      p{
+        text-align: justify;
+        margin: 0px 12px 12px 0px;
+        line-height: 1.8em;
+        padding-top: 12px;
+        min-height: 130px;
+        @media (min-width: 545px) {
+          display: flex;
+          align-items: center;
+        }
+      }
+  
+      h2 {
+        font-size: 28px;
+      }
+  
+      h2.signin{
+        max-width: 97%;
+        border-bottom: 0;
+      }
+  
+      h3{
+        color: #FD8006;
+        text-align: center;
+        margin-top: 10px;
+        font-size: 24px;
+      }
+  
+      .images-services{
+        width: 180px;
+        height: 120px;
+      }
+  
+      .App-element-start{
+        position: relative;
+        bottom: 260px;
+        @media (max-width: 1631px) {
+          bottom: 0px;
+          left: 145px;
+        }
+        @media (max-width: 1376px){
+          bottom: 260px;
+          left: 0px;
+        }
+      }
+    }
+  
+    .App-column-elements{
+      display: flex;
+      flex-direction: column;
+      margin-left: 18px;
+      margin-bottom: 18px;
+  
+      label{
+        font-size: 16px;
+        margin-bottom: 4px;
+      }
+      .App-element{
+        margin-bottom: 16px;
+      }
+  
+      .center-tools{
+        text-align: center;
+      }
+  
+      .center-text{
+        text-align: center;
+        max-width: 500px;
+      }
+  
+      .subtitle-home{
+        text-align: center;
+        font-size: 18px;
+        border-bottom: 0px;
+        @media (max-width: 768px) {
+          margin-left: 12px;
+        }
+        margin-left: 28px;
+        margin-top: 36px;
+      }
+  
+      div{
+        label{
+          max-width: 245px;
+          min-width: 245px;
+          display: inline-block;
+        }
+      }
+  
+    }
+  
+    .App-center-elements{
+      justify-content: center;
+      margin-right: 12px;
+      max-width: 500px;
+      display: inline-block;
+    }
+  
+    .App-font-event{
+      font-size: 1rem;
+    }
+  
+    .App-container_firstRow{
+      width: 100%;
+      border-bottom: 0.1em solid #E6E6E6;
+  
+      nav{
+        display: flex;
+        justify-content: space-around;
+        margin: 18px 0px 20px 0px;
+      }
+  
+    }
+}
+
+.App-link{
+  @media (max-width: 768px) {
+    padding: 8px 0px 4px 0px;
+  }
+  padding: 8px 8px 8px 8px;
+  font-size: 16px;
+  color: #FD8006;
+  border-radius: 5px;
+  transition: all 0.2s;
+  border: 1px solid #FD8006;
+  text-decoration: none;
+
+  i{
+    position: relative;
+    top: 5px;
+    left: 3px;
+  }
+
+  &:hover{
+    background-color: #FD8006;
+    color: white;
+  }
+}
+
+.App-active-link{
+  color: white;
+  background-color: #FD8006;
+  
+  &:hover{
+    text-decoration: underline;
+  }
+}
