+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/index.css
@@ -6,6 +6,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  height: 100%;
 }
 
 code {
